+++ packages/core/tailwind.build.config.js
@@ -1,0 +1,6 @@
+const preset = require("./tailwind-preset");
+
+module.exports = {
+  presets: [preset],
+  content: ["./src/styles/**/*.{scss,css}"],
+};